+++ CHANGES.rst
@@ -9,6 +9,8 @@ v14.0.0
 - Removed ``FastPath``. Just use ``Path``.
 - Removed ``path.CaseInsensitivePattern``. Instead
   use ``path.matchers.CaseInsensitive``.
+- Added ``Traversal`` class and support for customizing
+  the behavior of a ``Path.walk``.
 
 v13.3.0
 -------